Made data not persistent on docker compose down
Deleted the mount of mysql data folder on the volume, to make it not persistant on docker compose down.

GraphDB must stay mounted on the volume, to give python container access to the folders of GraphDB.
To deal with it, in the init container, on first launch, the folders are wiped.

+ Deleted the echo of not required to user passwords. They can be read in the config.ini file of python container.